services/media-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
minio==7.2.3
aiofiles==23.2.1
pydantic==2.5.3
pydantic-settings==2.1.0
